moved interfaces into datatypes
